Update the FFmpeg binaries loader system.
